Re #637 fix some minor bugs in parsing and parsing tests

+++ src/groovy/us/wearecurio/data/DataRetriever.groovy
@@ -411,7 +411,7 @@ class DataRetriever {
 		if ((!r) || (!r[0]))
 			return null
 			
-		return new TagUnitStatsInfo(userId, null, UnitGroup.get(unitGroupId), r[0]['unit'], r[0]['s'])
+		return new TagUnitStatsInfo(userId, null, UnitGroup.get(unitGroupId), r[0]['unit'], r[0]['s'].longValue())
 	}
 	
 }
